Linear search algorithm implementation

+++ search_algorithms/0-linear.c
@@ -9,7 +9,7 @@
 */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t counter = 0;
+	unsigned int counter = 0;
 
 	if (array)
 	{
